Completely reorganize codebase
