+++ mod.ts
@@ -1,1 +1,4 @@
 export * from "./src/CBOR.ts";
+export * from "./src/SimpleValue.ts"
+export * from "./src/TaggedValue.ts"
+export * from "./src/types.ts"
